# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/facilities/get_venues": {
      "get": {
        "tags": [
          "View Facility"
        ],
        "summary": "Get All Venues",
        "description": "Organizer or Audience can Access",
        "operationId": "get_all_venues_facilities_get_venues_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Show venues only by the Names.",
              "title": "Name"
            },
            "description": "Show venues only by the Names."
          },
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Show venues only from this location (can match partly).",
              "title": "Location"
            },
            "description": "Show venues only from this location (can match partly)."
          },
          {
            "name": "min_capacity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Show venues with at least this much capacity.",
              "title": "Min Capacity"
            },
            "description": "Show venues with at least this much capacity."
          },
          {
            "name": "max_price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Show only items with price less than or equal to this value.",
              "title": "Max Price"
            },
            "description": "Show only items with price less than or equal to this value."
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sort by this column name (like price, name, or capacity).",
              "title": "Sort By"
            },
            "description": "Sort by this column name (like price, name, or capacity)."
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "description": "Sort order — 'asc' for ascending, 'desc' for descending.",
              "default": "asc"
            },
            "description": "Sort order — 'asc' for ascending, 'desc' for descending."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/facilities/get_bands": {
      "get": {
        "tags": [
          "View Facility"
        ],
        "summary": "Get All Bands",
        "description": "Organizer or Audience can Access",
        "operationId": "get_all_bands_facilities_get_bands_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter bands by name.",
              "title": "Name"
            },
            "description": "Filter bands by name."
          },
          {
            "name": "genre",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter bands by genre.",
              "title": "Genre"
            },
            "description": "Filter bands by genre."
          },
          {
            "name": "member_count",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter bands by member count.",
              "title": "Member Count"
            },
            "description": "Filter bands by member count."
          },
          {
            "name": "max_price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Show only bands with price less than or equal to this value.",
              "title": "Max Price"
            },
            "description": "Show only bands with price less than or equal to this value."
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sort by this column name (like price, name, or member_count).",
              "title": "Sort By"
            },
            "description": "Sort by this column name (like price, name, or member_count)."
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "description": "Sort order — 'asc' or 'desc'.",
              "default": "asc"
            },
            "description": "Sort order — 'asc' or 'desc'."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/facilities/get_decorations": {
      "get": {
        "tags": [
          "View Facility"
        ],
        "summary": "Get All Decorations",
        "description": "Organizer or Audience can Access",
        "operationId": "get_all_decorations_facilities_get_decorations_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter decorations by name.",
              "title": "Name"
            },
            "description": "Filter decorations by name."
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter decorations by type.",
              "title": "Type"
            },
            "description": "Filter decorations by type."
          },
          {
            "name": "max_price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Show only decorations with price less than or equal to this value.",
              "title": "Max Price"
            },
            "description": "Show only decorations with price less than or equal to this value."
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sort by this column name (like price or name).",
              "title": "Sort By"
            },
            "description": "Sort by this column name (like price or name)."
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "description": "Sort order — 'asc' or 'desc'.",
              "default": "asc"
            },
            "description": "Sort order — 'asc' or 'desc'."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/facilities/get_snacks": {
      "get": {
        "tags": [
          "View Facility"
        ],
        "summary": "Get All Snacks",
        "description": "Organizer or Audience can Access",
        "operationId": "get_all_snacks_facilities_get_snacks_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter snacks by price.",
              "title": "Price"
            },
            "description": "Filter snacks by price."
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sort by this column name (like price or id).",
              "title": "Sort By"
            },
            "description": "Sort by this column name (like price or id)."
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "description": "Sort order — 'asc' or 'desc'.",
              "default": "asc"
            },
            "description": "Sort order — 'asc' or 'desc'."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/facilities/facility/": {
      "get": {
        "tags": [
          "View Facility"
        ],
        "summary": "Check Facility Available",
        "description": "Organizer or Audience can Access",
        "operationId": "check_facility_available_facilities_facility__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Facility ID",
              "title": "Id"
            },
            "description": "Facility ID"
          },
          {
            "name": "facility_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Name of the facility (venue/band/decoration)",
              "title": "Facility Name"
            },
            "description": "Name of the facility (venue/band/decoration)"
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Expected date of the facility (YYYY-MM-DD)",
              "title": "Date"
            },
            "description": "Expected date of the facility (YYYY-MM-DD)"
          },
          {
            "name": "slot",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SlotEnum",
              "description": "Slot of the facility (Morning/Afternoon/Night)"
            },
            "description": "Slot of the facility (Morning/Afternoon/Night)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/facilities/booked-events": {
      "get": {
        "tags": [
          "View Facility"
        ],
        "summary": "Get Booked Facility Events",
        "description": "Admin  Only  Access",
        "operationId": "get_booked_facility_events_facilities_booked_events_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "facility_type_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "Valid facility type ID",
              "title": "Facility Type Id"
            },
            "description": "Valid facility type ID"
          },
          {
            "name": "facility_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "Valid facility ID",
              "title": "Facility Id"
            },
            "description": "Valid facility ID"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Start date for filtering bookings",
              "title": "Start Date"
            },
            "description": "Start date for filtering bookings"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "End date for filtering bookings",
              "title": "End Date"
            },
            "description": "End date for filtering bookings"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SlotEnum": {
        "type": "string",
        "enum": [
          "Morning",
          "Afternoon",
          "Night"
        ],
        "title": "SlotEnum"
      },
...
      "SortOrder": {
        "type": "string",
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortOrder"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {

            },
            "tokenUrl": "users/login"
          }
        }
      }
    }
  }
}